pyproject.toml
[project]
name = "Blog_api_with_caching"
version = "0.2.3"
description = "Blog API with caching of posts"
authors = [
    {name = "Tychin Denis"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "django==6.0.3",
    "pillow==11.3.0",
    "gunicorn==25.1.0",
    "django-health-check==4.2.1",
    "redis==7.3.0",
    "djangorestframework==3.17.0",
    "django-cors-headers==4.9.0",
    "drf-standardized-errors==0.15.0",
    "drf-spectacular==0.29.0",
    "psycopg2-binary==2.9.11",
    "django-structlog==10.0.0",
    "structlog==25.5.0",
    "python-decouple==3.8",
    "django-query-counter==0.5.0"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "25.12.0"
ruff = "0.11.13"
django-debug-toolbar = "6.2.0"
pre-commit = "4.5.1"
model-bakery = "1.23.3"
pytest = "9.0.2"
pytest-django = "4.12.0"

[tool.black]
line-length = 99
target-version = ["py312"]
include = "\\.pyi?$"

[tool.ruff]
indent-width = 4
line-length = 99
required-version = ">=0.11.11"
target-version = "py312"
preview = true
exclude = [
    "src/alembic",
    "alembic",
    "data/docs_assets",
]

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 60
line-ending = "lf"

[tool.ruff.lint]
ignore = [
    "E265",
    "D100",
    "D105",
    "D104",
    "D203",
    "D211",
    "D213",
    "N818",
    "W505",
    "ANN401",
]
select = [
    "E3",
    "E5",
    "E4",
    "E7",
    "E9",
    "W",
    "C90",
    "I",
    "N",
    "D",
    "F",
    "ANN",
    "ASYNC",
    "A",
    "COM",
    "RET",
]

[tool.ruff.lint.extend-per-file-ignores]
"__init__.py" = ["E402", "E401"]

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"

[tool.ruff.lint.isort]
case-sensitive = true
known-local-folder = ["src"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.pycodestyle]
max-doc-length = 60
max-line-length = 99

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = ["test_*.py", "*_tests.py"]
testpaths = ["posts/tests"]
addopts = "-ra"
